+++ packages/feflow-cli/lib/core/plugin/loadPlugins.d.ts
@@ -1,1 +1,0 @@
-export default function loadPlugins(ctx: any): Promise<void>;