+++ packages/feflow-gui/src/renderer/components/create/create.vue
@@ -1,6 +1,6 @@
 <template>
   <div class="create-inner">
-    <section v-if="list.length">
+    <section v-if="generators.length">
       <el-form label-position="left" label-width="140px" ref="form" :model="formData">
         <el-form-item label="脚手架">
           <el-select v-model="targetGenerator" :disabled="isWorking" placeholder="请选择">
@@ -74,6 +74,9 @@ import { Terminal } from 'xterm'
 import { FitAddon } from 'xterm-addon-fit'
 import 'xterm/css/xterm.css'
 import SchemaForm from './schema-parser/components/vue-form'
+// import tKill from 'tree-kill'
+
+const TimeOutTsp = 20 * 10000
 
 export default {
   name: 'create-page',
@@ -90,7 +93,8 @@ export default {
       banner: '',
       hasInitTerminal: false,
       popoverVisible: false,
-      empty: false
+      empty: false,
+      watcherTimer: null
     }
   },
   mounted() {},
@@ -118,16 +122,35 @@ export default {
       if (!newValue) {
         this.popoverVisible = false
       }
+    },
+    configMap(newValue) {
+      this.getSchemaForm()
     }
   },
   methods: {
-    ...mapActions(['builConfig', 'getGenerator', 'selectWorkSpace', 'resetState', 'toggleWorkStatus']),
+    ...mapActions([
+      'builConfig',
+      'getGenerator',
+      'selectWorkSpace',
+      'resetState',
+      'toggleWorkStatus',
+      'resetGenerateList'
+    ]),
     init() {
+      // 重新获取
+      this.resetGenerateList()
       // 获取脚手架
       this.getGenerator()
-      this.resetState()
       this.toggleWorkStatus(false)
+      // 获取动态表单
+      this.getSchemaForm()
+    },
+    getSchemaForm() {
+      const _generatorsConfig = {}
+      const _generators = []
       const gens = Object.keys(this.configMap)
+
+      // 没有脚手架
       if (!this.list.length) {
         return (this.empty = true)
       }
@@ -137,15 +160,19 @@ export default {
           const key = genName
           const gen = this.configMap[genName]
 
-          this.generators.push({
+          _generators.push({
             value: key,
             label: gen.description
           })
 
-          this.generatorsConfig[key] = this.configMap[key]
+          _generatorsConfig[key] = this.configMap[key]
         })
+
+      this.generators = _generators
+      this.generatorsConfig = _generatorsConfig
+
       // 默认加载第一个脚手架
-      this.targetGenerator = this.list[0]
+      this.targetGenerator = (_generators[0] && _generators[0].value) || ''
     },
     // 初始化控制台
     initTerminal() {
@@ -217,17 +244,60 @@ export default {
       if (typeof childProcess === 'number') {
         return this.handleInitCode(childProcess)
       }
+
+      this.childProcessPid = childProcess.pid
+      this.childProcess = childProcess
+
+      // 清理上次的日志
+      this.term && this.term.clear()
+
       childProcess.stdout.on('data', data => {
+        this.steamWatcher()
         this.term.writeln(data)
       })
       childProcess.on('close', code => {
+        this.steamWatcherClose()
         this.handleInitCode(code)
       })
       childProcess.on('error', err => {
+        this.steamWatcherClose()
         this.messageInstance && this.messageInstance.close()
         this.messageInstance = null
         // 上报点
         this.toast('脚手架生成过程发生异常', err, 'error')
+      })
+    },
+    steamWatcher() {
+      if (watcherTimer) clearTimeout(watcherTimer)
+      watcherTimer = setTimeout(() => {
+        this.handleTimeout()
+      }, TimeOutTsp)
+    },
+    steamWatcherClose() {
+      if (watcherTimer) clearTimeout(watcherTimer)
+    },
+    handleTimeout() {
+      this.$alert('检测到当前任务耗时异常，是否中止？', '任务异常', {
+        confirmButtonText: '中止',
+        cancelButtonText: '再等等',
+        showCancelButton: true,
+        callback: action => {
+          // 拿到子进程pid然后kill
+          if (action === 'confirm') {
+            if (this.childProcess.exitCode === null) {
+              console.log('还未退出')
+              // 尝试退出子进程
+              this.childProcess.kill('SIGHUP')
+              this.messageInstance && this.messageInstance.close()
+              this.messageInstance = null
+              this.popoverVisible = false
+              this.toggleWorkStatus(false)
+            }
+          } else {
+            // 继续超时逻辑
+            this.steamWatcher()
+          }
+        }
       })
     },
     handleConsoleClick() {
@@ -293,6 +363,7 @@ export default {
       }
     },
     toast(title, msg, type = 'info', isPersistent = false) {
+      // TODO mixins
       let opt = {
         title,
         message: msg
@@ -307,10 +378,9 @@ export default {
       // 重置表单， 防止重复初始化项目
       this.toggleWorkStatus(false)
       this.banner = ''
-      this.$store.dispatch('Schema/init', { schema: this.targetGeneratorConfig })
-
-      // 重置脚手架状态
-      this.resetState()
+      if (this.targetGeneratorConfig) {
+        this.$store.dispatch('Schema/init', { schema: this.targetGeneratorConfig })
+      }
     }
   }
 }